added separate benchmarks file
